+++ tests/logger_test.go
@@ -14,7 +14,7 @@ func TestStructuredLogger(t *testing.T) {
 	t.Run("TextFormat", func(t *testing.T) {
 		buf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(buf)
 		l.SetFormat(logger.LogFormatText)
 		l.Info("hello %s", "world")
@@ -28,7 +28,7 @@ func TestStructuredLogger(t *testing.T) {
 	t.Run("JSONFormat", func(t *testing.T) {
 		buf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(buf)
 		l.SetFormat(logger.LogFormatJSON)
 		l.Info("hello %s", "world")
@@ -49,7 +49,7 @@ func TestStructuredLogger(t *testing.T) {
 	t.Run("WithFields", func(t *testing.T) {
 		buf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(buf)
 		l.SetFormat(logger.LogFormatJSON)
 		l2 := l.WithFields(map[string]any{"request_id": "123"})
@@ -68,7 +68,7 @@ func TestStructuredLogger(t *testing.T) {
 	t.Run("SQLJSON", func(t *testing.T) {
 		buf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(buf)
 		l.SetFormat(logger.LogFormatJSON)
 		l.SQL("SELECT * FROM users", time.Millisecond*10, "arg1", 1)
@@ -90,9 +90,9 @@ func TestStructuredLogger(t *testing.T) {
 		mainBuf := &bytes.Buffer{}
 		errorBuf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(mainBuf)
-		l.SetLevelOutput(logger.LogLevelError, errorBuf)
+		l.SetLevelOutput(logger.Error, errorBuf)
 
 		l.Info("this is info")
 		l.Error("this is error")
@@ -118,9 +118,9 @@ func TestStructuredLogger(t *testing.T) {
 	t.Run("LevelOutputOnly", func(t *testing.T) {
 		errorBuf := &bytes.Buffer{}
 		l := logger.NewStdLogger()
-		l.SetLevel(logger.LogLevelInfo)
+		l.SetLevel(logger.Info)
 		l.SetOutput(nil) // Disable default output
-		l.SetLevelOutput(logger.LogLevelError, errorBuf)
+		l.SetLevelOutput(logger.Error, errorBuf)
 
 		l.Info("this is info")
 		l.Error("this is error")