+++ tests/query_logger_test.go
@@ -19,7 +19,7 @@ func TestQueryStructuredLogging(t *testing.T) {
 
 	buf := &bytes.Buffer{}
 	l := logger.NewStdLogger()
-	l.SetLevel(logger.LogLevelInfo)
+	l.SetLevel(logger.Info)
 	l.SetOutput(buf)
 	l.SetFormat(logger.LogFormatJSON)
 	db.SetLogger(l)